security for supermarkets

+++ app/services/product_manager/destroyer.rb
@@ -1,9 +1,10 @@
 module ProductManager
   class Destroyer
-    attr_reader :product_id
+    attr_reader :product_id, :supermarket
 
-    def initialize(product_id)
+    def initialize(product_id, supermarket)
       @product_id = product_id
+      @supermarket = supermarket
     end
 
     def call
@@ -23,9 +24,12 @@ module ProductManager
     end
 
     def scope
-      @product = Product.find(product_id)
+      @product = supermarket.products.find_by(id: product_id)
+      if @product.nil?
+        raise ActiveRecord::RecordNotFound, "Product not found for this supermarket"
+      end
       unless @product.destroy
-        raise StandardError.new(product.errors.full_messages.to_sentence)
+        raise StandardError.new(@product.errors.full_messages.to_sentence)
       end
     end
   end
